Add support for Postgres
This adds support for the `PostgreSql` `DatabaseType`, by reusing most
of the code that was added for `MsSql`. To do this it uses ADO.NET
provided interfaces, instead of specific types provided by
`Microsoft.Sql`. It also creates an `IQueryBuilder` interface, because
this is different between PG and SQL, but most of the other things in
SQLQueryEngine, can be shared.