Improve configuration managament
This changes our config handling, such that we can use the standardized
ASPNETCORE_ENVIRONMENT environment variable to switch between
environments. It also adds a default/sample postgres and mssql
configuration.

The first reason for this change is that it allows developers to easily
switch between different configurations. This can be useful if they need
to test multiple databases or multiple schemas.

The second reason for this change is that people might need a custom
connection string for local testing, and this connection string is only
applicable to them, e.g. it contains a password. Without this change
they would have to modify the `appsettings.json` file that's checked in
to git. This would result in a dirty working directory and they might
accidentally commit this local change.

This change allows these people to have a custom config file called
`appsettings.MsSqlCustom.json` that's not checked in. That file can
safely contain their custom connection string.