Add Query Resolver implementation for MsSql (based on predefined config files) (#21)
## Purpose
Adds implementation of a simple JSON config file based GraphQL query resolver for Microsoft Azure SQL DB.

## Why this change is needed
Builds a GraphQL POC for SQL extending from the Cosmos POC using HotChocolate

## How
 By adding generalizations for
- db connection service,
- query engine, 
- mutation engine,
- query executor
The schema and resolver definitions are read from the file in the constructor of GraphQLService so that after running the app, we can simply run GraphQL queries against SQL.

Please reference this updated Class diagram to better understand all associations between classes followed by the brief description of the changes.
#### Class Diagram: SqlQueryEngine
![SqlQueryEngine](https://user-images.githubusercontent.com/3513779/133938020-6c6aaea6-fb68-4e59-abd0-15da59b526c0.PNG)

#### Class Diagram: CosmosQueryEngine
![CosmosQueryEngine](https://user-images.githubusercontent.com/3513779/133938040-5504e48e-f914-407e-ba6a-a192755a07a2.PNG)

#### ConfigurationProvider class
Reads configuration of the app from appsettings.json and maintains a DbType property to distinguish which DB the app targets.

#### Startup class
Based on the DbType property in ConfigurationProvider, instantiates the required services for the respective DB.

#### IDbConnectionService interface
Derives from the IClientProvider interface and is required for Sql like databases(e.g. Sql db, Postgres) to get an open connection before executing a query.

#### MsSqlClientProvider class
Implements the IDbConnectionService interface so that an open connection to SQL DB can be retrieved. 
It uses the MsSql credential properties provided by the ConfigurationProvider to generate a connection string if not already specified in appsettings.json

#### IQueryEngine interface
Every database should implement this interface for execution of GraphQL queries against that database.

#### CosmosQueryEngine class
Implements the methods to execute GraphQL queries against a Cosmos DB.

#### SqlQueryEngine class
Provides methods to translate GraphQL queries to SQL queries against a SQL like DB and uses the QueryExecutor to run those against that DB.

#### IQueryExecutor interface
Provides interface for execution of queries against a Sql like DB.

#### QueryExecutor class
Uses an implementation of IDbConnectionService to get an open connection and implements async Execute methods against a Sql like database.

#### IMutationEngine interface
Every database should implement this interface for execution of GraphQL mutations against that database.

#### CosmosMutationEngine
Implements the methods to execute GraphQL mutations against a Cosmos DB.

#### SqlMutationEngine
Not yet implemented but added because the GraphQLService depends on having an instance of this class at runtime.

### Test

- Made sure existing tests against Cosmos Db work as expected
- Modified appsettings.json, schema.gql and config.json to point to an existing SQL DB and verified the GraphQL queries return data as expected. 

#### appsettings.json
```
  "DatabaseConnection": {
    "DatabaseType": "MsSql",
    "Credentials": {
   ...
       "Server": "localhost",
      "ConnectionString": ""
    }
```
With the sample schema.gql file, here is a sample 
#### config.json
```
{
  "GraphQLSchema": "",
  "Resolvers": [
    {
        "engine": "SQL",
        "id": "characterById",
        "databaseName": "StarWars",
        "isList": false,
        "parametrizedQuery": "SELECT [id], [name], [type], [homePlanet], [primaryFunction] FROM character WHERE [id] = @id"
    },
    {
        "engine": "SQL",
        "id": "characterList",
        "databaseName": "StarWars",
        "parametrizedQuery": "SELECT [id], [name], [type], [homePlanet], [primaryFunction] FROM character"
    },
    {
        "engine": "SQL",
        "id": "planetById",
        "databaseName": "StarWars",
        "parametrizedQuery": "SELECT [id], [name] FROM planet WHERE [id] = @id"
    },
    {
        "engine": "SQL",
        "id": "planetList",
        "databaseName": "StarWars",
        "parametrizedQuery": "SELECT [id], [name] FROM planet"
    }
  ]
}
```

### Known Issues:
Int values with nulls in the database are returned with errors.