Make dotnet watch run restart on json and gql changes
When running with `dotnet watch run` the binary gets rebuilt and
restarted whenever a `.cs` file changes. For this project a lot of
behaviour is defined in the config files though. This makes sure the
`watch run` restarts the program whenever a json or gql file is modified
in the project.