Disable 'Just My Code' debugging in vscode
We use HotChocolate in an uncommon way, this means documentation is
often of little help. So instead when we run into issues we need to read
the source code. Finding the right piece of code to read can be made a
lot easier if the debugger can show us what code is being executed. By
disabling 'Just My Code' the debugger will allow us to debug code from
the HotChocolate library that's being executed.