Update Pipeline to create and push docker image (#61)
# Why is this change being made?
So we can automatically generate docker images and push them to a container registry.

# What changed?

- Updated the pipeline to have: 
    - A task to get the build start time for the tag of the docker image.
    - A buildAndPush task which creates and pushes an image to docker hub. We're currently using my private docker hub while we figure out where we want these images to be pushed.

- Updated the documentation to fix some errors and added some information on how to use these new docker images.
- Changed DockerFile copy path from Debug to Release since the pipeline builds release.
Not in this PR:
- Added a service connection so the Azure pipeline can communicate with the dockerhub from this pipeline config.

# How was this validated?
Created a duplicate pipeline for my fork of this repo. I then validated by creating a commit that the image ended up on the docker hub and followed the instructions that I added to validate that the container could be run.