Reenable logging stacktraces of uncaught exceptions
Logging of stacktraces was disabled by #97. Based on PR comments the
reason for this was that for most types of errors `error.Exception`
would be `null`. Thus the exception logging code would actually create a
new error by dereferencing a `null`.

This was addressed by only logging the error `Code` and `Message`.
However, logging these is not very useful. They are already passed to
the frontend. So this fixes the problem in a better way, by only logging
to stderr when the error its `error.Exception` field is non-null.