Fix git commit hook after directory refactor
In [#65][1] directory structure was simplified. The git commit hook had
some code that was required to address this more complex directory
structure. After the directory structure simplification this now
unnecessary code had the unintended effect of the causing the Test
directory not to be formatted by the hook. This fixes the commit hook.

[1]: https://github.com/Azure/hawaii-gql/pull/65