Remove write functionality from IMetadataStoreProvider
We currently only have one IMetadataStoreProvider implementation (except
for testing) and it doesn't support storing data. This removes the store
functionality from the interface. The store functionality was used
previously when the metadata was stored in the backing datastore.

It's possible that we'll support storing metadata in the database in the
future again. Instead of letting this unused code rot away until we do
so, I think it makes more sense to remove it for now. Then when we add
this functionality again later, we can design an API at that point.