MySql Integration
MySql should now have full MsSql and PG capabilities, but there are some issues that need to be addressed ideally by someone with more knowledge on MySql:
* MySql provides no RETURNING or Output keywords so I use a double query for insertions and updates. Maybe due to this, insertions and updates for MySql
run significantly slower than those for PG and MsSql
* The suggested query generation pattern for MySql is slightly differet as far as query arrangement goes, but a big difference is the introduction
of `JSON_EXTRACT`(jsonString, '$') to convert nested results of `JSON_ARRAYAGG` to JSON when used as a parameter of `JSON OBJECT`. e.g. `JSON_OBJECT("name": JSON_ARRAyAGG(
... ))` would result in `{"name": "[...]"}` instead of `{"name": [...]}`.
* `JSON_ARRAYAGG` which is widely used to produce JSON arrays does not guarantee result order. In the tests so far, the outputs obey the order of order by, but that is
not guaranteed. I noticed how in one particular query, if I deleted `LIMIT 100` the elements came out unordered. This is something that needs to be addressed.