Improve REST insert error message to contain bad fields (#173)
Without this the error is not actionable for a user of the API. Because
there's no indication of what they did wrong.

There's a question of leaking internals, but at this point we're not
leaking much. We might want to revert/change this behaviour in the
future to limit leaking.