Mandate Fields section in runtime config only if custom fields are present (#195)
For GraphQL types, that don't have custom fields, we don't need a "Fields" section in the runtime config file.
We only need that section for custom fields because it is only for those fields that we need additional information e.g. relationship type, pagination type etc.
This PR fixes the unnecessary mandate we have today when there are no custom fields.