Fix PUT request validation and null handling (#240)
This change fixes the following issues in PUT and adds corresponding regression tests:

In request validation, we are forcing nullable columns to be present in the request body if they are not autogenerated. Change the following condition of validity from
!column.Value.IsAutoGenerated && fieldsInRequestBody.Contains(column.Key) to (!column.Value.IsAutoGenerated && column.Value.IsNullable)
Added PutOne_Insert_Nullable_Test to test this

In request validation, we are forcing non-nullable non-primary key fields to be present in request body even if they are autogenerated. Fix by changing validity from
!column.Value.IsNullable && fieldsInRequestBody.Contains(column.Key)
to !column.Value.IsAutoGenerated && !column.Value.IsNullable && fieldsInRequestBody.Contains(column.Key)
Added PutOne_Insert_AutoGenNonPK_Test as +ve
Added PutOne_Insert_BadReq_AutoGen_NonNullable_Test as -ve test

For adding unspecified fields in request body with null values, we are erroring out with Request must define values for all nullable and non-default valued columns. Fix this by adding a continue :
// If the left over column is autogenerated or has a default, // then no need to add it with a null value. if (tableDefinition.Columns[leftOverColumn].IsAutoGenerated || tableDefinition.Columns[leftOverColumn].HasDefault)
Added PutOne_Insert_AutoGenNonPK_Test

Use IsDBNull to check result of dataReader and construct the returning JSON object with null value.
Without this fix, the JSON constructed is incorrect :
Today, for example, for this request :
PUT https://localhost:5001/magazines/id/3 with body
{ "title" : "People" }
the result is:
{ "id": 3, "title": "People", "issueNumber": {} }
With this fix, it is:
{ "id": 3, "title": "People", "issueNumber": null }

Fix sql-config.json to correct isNullable field.