Improve user feedback on db query failure (#318)
Addresses #152

Reports generic db errors instead of an unexpected server error when queries fail due to errors in the database.

MsSql also does not throw the exception during `ExecuteReaderAsync`, but during reading with `ReadAsync` from the `DbDataReader`. Since adding try catch blocks whenever `ReadAsync` is used feels unnecessarily verbose, I created a wrapper method for `ReadAsync` which I put into the `QueryExecutor` where `DbException`-s are caught, and that wrapper method is used instead of calling `ReadAsync` on the `DbDataReader` directly.

Different error codes are thrown for each database. I am working on UPSERT support for Pg, and that is likely to throw a different error for the REST Api tests under this implementation of `DbException` handling. To solve this, I preemptively introduced an `expectedExceptionMap` to the REST Api tests, to account for testing for different errors for different databases. If it is this is not the case and a different exception will not be thrown by Pg, I will remove this particular change.