Add One-To-One Join Support for GraphQL (#320)
Addresses #127 

Adds support for One-To-One relationships.
In addition to One-To-One support, this also changes the `ForeignKeyDefinition` to bring it closer to actual db foreign keys.
Each foreign key can have `columns` and `referencedColumns` much like actual foreign keys. These properties are not mandatory and if they are not available the table and referenced table pk-s will be used as columns and referenced columns respectively.