Read the new developer friendlier runtime config file format and store for later consumption (#338)
Summary
We are changing from using a pre-set configuration file, to one that is created based on the inputs provided by a developer using a command line tool. Once generated this config file then needs to be read into the appropriate objects for later consumption.

This work seeks to read and deserialize all of the relevant properties in this new developer friendlier runtime configuration file format into the RuntimeConfig class that we create for this PR.

The following RFC was used as a guide for determining what properties will need to be consumed, and an example runtime-config-test.json was generated for testing based on this RFC:
https://github.com/Azure/project-hawaii/blob/main/rfcs/configuration-file.md

A simple test was added that will instantiate a RuntimeConfig based on the example runtime-config-test.json that has been created, and then verify that deserialization happens correctly.