Patch #379 since the config was not valid (#391)
This was my mistake.
I did not rerun the application after I did the merge with main, I only ran the tests.
If I had run the application, I would have seen that the config was not valid and the run is unsuccessful.

This PR fixes the config, but also adds a test to check if the config is valid or not so this scenario does not repeat again.

The issue was that due to change in GraphQLBuilder.FormatNameForObject function I switched to GraphQLBuilder.FormatNameForField which does not capitalize the first letters of the name. That was causing the config validator to fail.
I decided to remove this completely since the gql type capitalization (whatever was decided by the schema generation or the user directly) should be appropriate to use to name [TypeName]FilterInput and [TypeName]OrderByInput input types without any extra formatting.