implementing default value support Adding a new directive, defaultValue which uses a oneOf directive'ed input type to indicate which field to pick the default value from
Might not need to unpack quite the same here, as oneOf should do the unpacking at the GraphQL end, but then we need to generate the field type on the input slightly differently, as it'll need to be a DefaultValue type not the underlying scalar.
It's worth noting that oneOf is still in draft at time of implementation (https://github.com/graphql/graphql-spec/pull/825) but has made it to the main spec branch.