Ignoring graphQLschema from CosmosDBOptions in generated config Json (#541)
# What is the Issue?
```
public record CosmosDbOptions(

        string Database,

        string? Container,

        [property: JsonPropertyName(CosmosDbOptions.GRAPHQL_SCHEMA_PATH_PROPERTY_NAME)]

        string? GraphQLSchemaPath,

        string? GraphQLSchema)

    {

        public const string GRAPHQL_SCHEMA_PATH_PROPERTY_NAME = "schema";

        public const string JSON_PROPERTY_NAME = nameof(DatabaseType.cosmos);

    }
```



Generated Config:

```
"cosmos": {

    "database": "graphqldb",

    "container": "planet",

    "schema": "schema.gql",

    "graphqlschema": "content of schema.gql"

  }
```
json property name of GraphQLSchemaPath is "schema". So, while generating config under cosmosDB we are getting schema which graphQLSchemaPath and one graphQlSchema which is confusing.

graphQLSchema is for the runtime and cannot be specified before starting of the engine, therefore, redundant for the config.

# What is the Fix?
we will remove graphQLSchema from config, i.e. it won't be visible in the config, but still be present in the properties of CosmosDbOptions because it is used by engine during runtime. So, we can use JsonIgnore to not serialize it while generating the config.

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>