Fix sample config for CORS and docker README (#573)
## What is the issue?
Sample docker-compose files specify port 5000. However, in the configs, we enable port 3000 in the cors section. 
This causes requests sent to the container set up using the default docker compose files to fail in CORS policy execution.

## What is the change?
This is to sync the docker compose and configuration files and the chosen port is 5000. Also, note for container scenario, we don't have `https`. 
It also fixes the `GetStarted.md` to include Banana Cake Pop settings to make sure the default configuration of schema endpoint is accurate. In my case, it was set to `http://localhost:5000` instead of `http://localhost:5000/graphql`. 

## Testing
- Built a docker image using the GetStarted readme and made sure I was able to query both Cosmos and SQL.
- Also pulled the latest image from the private container registry to perform the same checks.