Wildcard action support in config (#577)
This change is required to support wildcard (*) action in the config file for an enitity role combination. Previously, we used to get this error when we had wildcard as an action:

System.Private.CoreLib.dll!System.ThrowHelper.ThrowKeyNotFoundException<System.__Canon>(System.__Canon key) Line 213	C#
 	System.Private.CoreLib.dll!System.Collections.Generic.Dictionary<string, System.Collections.Generic.List<string>>.this[string].get(string key) Line 199	C#
 	Azure.DataGateway.Service.dll!Azure.DataGateway.Service.Authorization.AuthorizationResolver.SetEntityPermissionMap(Azure.DataGateway.Config.RuntimeConfig runtimeConfig) Line 301	C#
 	Azure.DataGateway.Service.dll!Azure.DataGateway.Service.Authorization.AuthorizationResolver.AuthorizationResolver(Azure.DataGateway.Service.Configurations.RuntimeConfigProvider runtimeConfigProvider, Azure.DataGateway.Service.Services.ISqlMetadataProvider sqlMetadataProvider) Line 45	C#


How was this done?
Whenever we get a wildcard action for an entity/role combination, we translate it into the 4 allowed CRUD operations i.e. create,read,update,delete. While setting the AuthorizationResolver.EntityPermissionMap, we include mappings for all the 4 operations instead of the wildcard operation.

How was this tested?
Included negative/positive tests for different authorization stages in AuthorizationResolverUnitTests.cs file where the wildcard is present as an action.