Upload build output for each CI pipeline (#656)
This change package engine and CLI in different platform explicitly so that it generates executable for each platform. THe binaries generated is a self-contained binaries. https://docs.microsoft.com/en-us/dotnet/core/deploying/deploy-with-cli#self-contained-deployment
This is done because the user might not have .NET runtime installed on their machine. So we would like reduce the adoption friction by bundle the runtime with the binaries.

The binaries are zipped so the user can just download the zip file, extract and run the CLI.

I also consolidate the build output to ./out directory. Under this directory, I separated the output to engine, cli and tests.