simplification of the arguments object graph being passed around
Rather than a partially-parsed GraphQL object graph, we now create a basic tree of nested dictionaries. We are stuck with object? around since things can be of a variety of type, but short of creating our own data structures, it's ok (just requires a bit of type casting).

This also means the methods for generating predicates should be a lot simpler, since they don't also have to do GraphQL parsing