Add version to the binaries built (#688)
### Issues
**Issue 1**
I noticed that the binaries generated from the pipeline does not have the correct version information. 
![image](https://user-images.githubusercontent.com/922655/184090694-0e7f7e16-2b4d-4530-a45e-77674a85db35.png)
**Issue 2**
I also notice that we generate a nuget with the build id suffix. This actually means that the nuget file is a prerelease. For releasing, we should not generate a version number with the suffix.
https://docs.microsoft.com/en-us/nuget/concepts/package-versioning#pre-release-versions

### Fixes
**Issue 1**
Passing version number to the build command and publish command. Version is passed to the build command so that the binaries built will have a correct version. Same reason for the publish command with the addition that the zip file will have version in the file name as well.
For the image below, I download the file from the build pipeline artifact.
![image](https://user-images.githubusercontent.com/922655/184092770-237456b7-70d6-485d-a2d5-e3b0bc8f6882.png)

**Issue 2**
I added a condition in the build pipeline to create a version with build id suffix if isNugetRelease variable is not set.

This is a run with isNugetRelease = true
https://msdata.visualstudio.com/CosmosDB/_build/results?buildId=69419696&view=results
The nuget package is uploaded with correct version
https://msdata.visualstudio.com/CosmosDB/_artifacts/feed/DataApiBuilder/NuGet/dab/versions/0.0.3-69415575
![image](https://user-images.githubusercontent.com/922655/184091774-d973d7bb-4cf1-4870-9998-7954dc2fb8c3.png)

This is a run with isNugetRelease = false
https://msdata.visualstudio.com/CosmosDB/_build/results?buildId=69419687&view=results
![image](https://user-images.githubusercontent.com/922655/184091703-a390e36b-47e0-4771-a230-96ca9a88e323.png)

**Additional changes**
- Add build output directory as argument to publish script so we can configure publish output from the pipeline.

I separate the counter for the **patch** version between PR runs and CI runs (and other [build reasons](https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables?view=azure-devops&tabs=yaml#build-variables-devops-services) like 'Manual'). This is done because we have lot of PR build and it will bump up a lot of patch version. This way we can keep the patch version number low. However, because we are manually triggering the release, the 'Manual' build reason is the one controlling the patch version number right now.