Code signing the binaries (#778)
Why make this change?
Closes #712
Closes #741

This change is needed to sign the binaries that we release with Microsoft digital signature. This is required for any product that we will be releasing publicly or privately from Microsoft.

What is this change?
This change added a template for code signing related steps. Because we are doing the signing, we have to split the nuget package and zip file creation steps. The flow of the pipeline is now

Build - I have to update the build step to use dotnet pack which implicitly will do the build. So that it will generate the binaries in the correct location for packing the nuget package later in step 3 after signing.
Sign the binaries that will be packed inside the nuget package
Create a nuget package
Sign the nuget package
Generate binaries for each platform (win, linux, osx)
Sign the binaries generate in step 5
Zip the signed binaries.
Generate SBOM
Upload the output to pipeline artifacts.
In code signing, we split the steps into 3 parts because they use different signing key code.

Sign binaries owned by Microsoft (our code).
Sign Third party binaries that we depend on.
Sign nuget package.
Additionally, I remove publish.sh since we can use powershell version. There is no need to keep 2 version of the scripts.
I also rename the zip file that we created to dab.zip from cli.zip.

How was this tested?
I scheduled a build and verified that the binaries and nuget package are signed.
https://msdata.visualstudio.com/CosmosDB/_build/results?buildId=70980233&view=results