Bug fix for OrderBy with int types and null values (#937)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/934

We were returning an Unexpected Error when attempting to use `orderby` along with `int` type columns that had `null` values.

## What is this change?

In `SqlPaginationUtil` we account for `null` values within the `JsonElement` we attempt to resolve to a scalar value. Now passing back `null` in this case

![image](https://user-images.githubusercontent.com/93220300/199373627-4e2b630b-eab3-44e6-b90c-0c320cb8d8ff.png)


In `SqlQueryStructure` we account for creating a param name when the value is now `null` due to the above change.

![image](https://user-images.githubusercontent.com/93220300/199373559-ab3bcdea-45ac-4148-afb2-b4491adfdd0f.png)


In `BaseSqlQueryBuilder` we need to account for the `null` values in a `nextLink` so we modify `MakePaginationInequality` so that is can accomodate `null` values, which will need to be set as `is NULL` rather than something like `= paramN`

![image](https://user-images.githubusercontent.com/93220300/199373508-2e4bd8c2-acb2-4261-8055-4e863029e2be.png)



## How was this tested?

We add a test that uses the `type_table` which includes integer types with null values. More testing will come in a future PR.

## Sample Request(s)

The bug was discovered with the following request
`https://localhost:5001/api/books?$first=5&$orderby=pages`

this can easily be modified to work with our current configuration and database setup if you modify the `Book` entity to have a column that is `int` type, which can be done with the following SQL

`ALTER TABLE dbo.books
ADD pages int;`

You can then repro the bug with the following get request
`https://localhost:5001/api/Book?$first=5&$orderby=pages`