Log Query Information (#1141)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/972

We want to assist in debugging and helping users to understand how the service is working. A part of that is providing all of the queries that are generated as a part of logging when the service is launched not in a hosted scenario at the log level of debug.

## What is this change?

We remove the previous logging of queries. Specifically, the foreign key query information that was being logged, and instead log all `sqltext` used by the `queryexecutor`.

### Before this change
We log only Foreign Key query information from directly within the query builders

### After this change
We log all query information from the query executor.

## How was this tested?

Testing can be done by simply running any request and watching the generated query in the log information, so long as the service was started at or beyond log level debug.

## Sample Request(s)

Any request of choice will work.