Fixing problem where duplicate columns are added to Cosmos queries
This can happen because it is valid to specify a field identifier multiple times. Most common scenarion is when it comes to unpacking a fragment, as the field might already be requested, but there's nothing illegal about repeating a field in general.

Added an explicit test, as it was only implicit through the fragment tests.

So far only Cosmos is tested/implemented