Fix include fields section behaviour in the config (#1282)
## Why make this change?
Fixes: https://github.com/Azure/data-api-builder/issues/1208.

## Need to see some test cases first!
What are we going to conclude from below examples:

**If** `include` **is not specified: it will resolve to all the fields present**
**If** `exclude` **is not specified, it will resolve to an empty array**

Consider the following cases for the fields section for a particular role/action combination for an entity:

1. `"fields":{
     "include": []
      }`

   Since include is explicitly specified as an empty array, we won't resolve it to wildcard. (We were doing this before during config validation but not during authorization resolution stage). **No field will be accessible for this configuration.**

2.  `"fields":{
     "include": [],
     "exclude":[]
      }`
Again since include/exclude both are empty arrays- **behavior is identical to first case.**

3. `"fields":{
      }
`
In this case, included/excluded are not declared at all, and both would assume their default values, i.e., include will resolve to all the fields present in the entity and exclude will be empty. **All the fields will be accessible in this configuration**.

4. `"fields":{
      "excluded": ["some_field"]
      }
`
In this case, included is not declared at all, so, it would assume its default value, i.e., it will resolve to all the fields present in the entity. **All the fields except '_some_field_' will be accessible in this configuration**. This is the very case which was mentioned in the linked issue https://github.com/Azure/data-api-builder/issues/1208.


## What is this change?
`include` is resolved to contain all fields only when it is null (not declared in the config). The behavior was the same even before during authorization resolution stage, but during config validation, we were missing this resolution.

## How was this tested?
- [x] Unit Tests added to `ConfigValidationUnitTests.cs` class.