Dispose of GraphQL Json Document Result (#1292)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/821

`JsonDocument` implements `IDisposable`,  and therefore needs to have `Dispose()` explicitly called for each object that is created. For more information on the `IDisposable` interface see: https://learn.microsoft.com/en-us/dotnet/api/system.idisposable?view=net-7.0

To prevent memory leaks, we dispose of the `JsonDocument` objects created as a part of the GraphQL results.

## What is this change?

- Where possible we use the `using` clause
-  otherwise, we clone the root element of any json documents that are created as a result of invoking the GraphQL middleware, and then dispose of the original json document object.

## How was this tested?

We run against the test suite, and this was manually tested with a variety of GQL requests.

## Sample Request(s)

Any valid graphQL request should work. One such example

`{
  books
  {
    items {
      title
    }
    hasNextPage
    endCursor
  }
}`