Reenable build trigger even for docs PRs (#1350)
## Why make this change?

By excluding triggers for PR changes in the `docs` folder with #1345, my
expectation was that the required statuses will be skipped for docs
changes. However, this was not the case, the builds were no longer
triggered but the statuses were still required as seen here:


![image](https://user-images.githubusercontent.com/3513779/227005408-5c40285b-8b2b-4390-97eb-bbdbfe451995.png)

Hence, no longer excluding docs changes from triggering builds. 

## What is the change?

Reenable `docs` PR triggering builds, so that the required statuses
continue to be checked.

### Note
There is no setting at the branch level to selectively define required
statuses. Until that is supported, we will require builds even for docs
PRs.