Fix serialization issue for config (#1386)
## Why make this change?

- Closes #1321
  - special character like '&' is escaped during serialization.

## What is this change?

- Updated the Encoder used in the config serialization.
- Now using
[`UnsafeRelaxedJsonEscaping`](https://learn.microsoft.com/en-us/dotnet/api/system.text.encodings.web.javascriptencoder.unsaferelaxedjsonescaping?view=net-7.0#system-text-encodings-web-javascriptencoder-unsaferelaxedjsonescaping),
it does not escape HTML-sensitive characters such as <, >, &. Since we
do not want any kind of change during string serialization, it's the
best we can use.
- It also allows
[UnicodeRanges.All](https://learn.microsoft.com/en-us/dotnet/api/system.text.unicode.unicoderanges.all?view=net-7.0#system-text-unicode-unicoderanges-all)
to go through unescaped. Hence, it doesn't regress our existing changes.

## How was this tested?
- [x] Unit Tests
- I created a new user in Azure SQL with a password containing special
characters such as [!@#$%^&*() ], and i was able to start the engine
succesfully.

## NOTE:
Since the config file is json and it is storing connection-string as
string,there are still some characters that can't be used like single
quotes `'` or double quotes `"` , or a backslash `\` as password.