Return generic error messages in production mode (#1357)
## Why make this change?

- Closes #671 
    - Consider a GET request such as : `api/Book/id/one`. 
The parameter type for `id` should be an int. In such a case, the error
message returned is the following
`"Parameter \"one\" cannot be resolved as column \"id\" with type
\"Int32\".",`
While this error message is helpful, it reveals more information than
necessary. For example, it can be deduced from the error message that
the database object backing the entity `Book` is a table and the system
type for the column `id` is `Int32`
- To not reveal additional information, a generic error message:
`Invalid value provided for field: id` will be returned when running in
`Production` mode. When running in `Development` mode, the detailed
error message will be returned.
- When parameters/arguements are provided with incorrect type values in
the graphQL requests, a generic error message : `The specified argument
value does not match the argument type` is returned irrespective of the
host mode being Development or Production.

## What is this change?

- A new method `IsDevelopmentMode()` is introduced in the
`ISqlMetadataProvider` interface and implementations for this method is
provided by the implementing classes.
- Exceptions are constructed with different error messages for
`Production` and `Development` host modes.
- Tests are added only for SQL database types as the error messages for
REST requests reveal additional information about the entities. GraphQL
requests return a generic error when running in both Development and
Production mode. So, no tests are added for Cosmos.

## Additional Refactoring Changes
- A common method `GetParamAsSystemType` is introduced which casts the
parameter values for tables, views and stored-procedures to the
respective column/parameter system types. Separate methods for such
casting for tables, views and stored-procedures is refactored to a
single common method.

## How was this tested?

- [x] Integration Tests
- [x] Manually

## Sample requests

Error Message when running in `Development` mode:


![image](https://user-images.githubusercontent.com/11196553/227622481-ccd0ce42-ce79-4b10-9270-f83af27313a5.png)

Error Message when running in `Production` mode:


![image](https://user-images.githubusercontent.com/11196553/227622698-f4e165e5-7567-42cb-ba86-f8cd5899581d.png)