Validate that fields referenced in db policy for create operation are present in request body (#1391)
## Why make this change?
Fix https://github.com/Azure/data-api-builder/issues/1388

## What is this change?
A set of fields (`FieldsReferencedInDbPolicyForCreateAction`) is added
to `BaseSqlQueryStructure` class which will store all the fields
referenced in the db policy for create action. As we add columns to the
`InsertColumns` list for `SqlInsertQueryStructure` (PUT/PATCH/POST
operations) one by one, we remove the columns if present from the
`FieldsReferencedInDbPolicyForCreateAction`. So, at the end if the size
of this set `FieldsReferencedInDbPolicyForCreateAction` is 0, it means
that all the fields referenced in the database policy for create action
are being included in the insert statement, and we are good. However, if
the size is non-zero, we throw a bad request exception.

## How was this tested?

- [x] Integration Tests

## Sample Request(s)

**Operation Type**: POST
**Api**: REST
**URL**: https://localhost:5001/api/Revenue/
**Database policy configured for operation:** "@item.revenue gt 1000"
**Reason why we will get an exception:** The field `revenue` referenced
in the database policy is not one amongst the fields that are being
inserted.

Request body: 
```json
{
    "id": 18,
    "category":"book",
    "accessible_role": "Anonymous"
}
```

Response:
```json
{
    "error": {
        "code": "BadRequest",
        "message": "One or more fields referenced by the database policy are not present in the request body.",
        "status": 400
    }
}
```