Improved stored procedure parameter value type handling during GraphQL schema creation (#1383)
## Why make this change?

- Closes #1221 which describes how the GraphQL schema fields created for
stored procedures do not use the proper value types for stored procedure
parameters. This issue occurs because the GraphQL schema builder was
trying to infer the value type of a parameter default value defined in
the runtime configuration. Runtime configuration is not enough to know
the target value type for casting. Database metadata is required.

## What is this change?

- In the GraphQL query and mutation schema builders, a databaseobject is
now referenced for determining the value type for which to cast the
runtime configuration defined default parameter value.

## How was this tested?

- [ ] Integration Tests
- [x] Unit Tests - Unit tests which cover the affected GraphQL schema
builder methods

## Sample Request(s)

- The example config and SQL scripts from issue #1221 can be used to
confirm locally. The value types of parameters in the database schema
are represented as supported graphQL value types.