Introducing parameters' DbType before passing on to the database - MsSql (#1442)
## Why make this change?
Fixes https://github.com/Azure/data-api-builder/issues/475
Most column types will accept a string null and the db will
automatically infer our intention, but for varbinary in MsSql the db
complains that it cannot convert varchar to varbinary. For more details
about the issue refer:
https://stackoverflow.com/questions/29254690/why-does-dbnull-value-require-a-proper-sqldbtype

## What is this change?
Changing how we form the `parameters` as we pass them to the database.
Previously we used to only pass the parameter's value. Now, we would
also pass the parameter's `DbType` along with the value as suggested in
the stack overflow article.

**Notes:**
1. We would only be passing the DbType for MsSql.
2. Usually, we haven't seen any need to pass the `DbType` for Sql Server
data types other than `varbinary`. However, we will still be supplying
the DbType for all other parameter's for which we are confirmed about
the 1:1 mapping from SystemType to DbType. Sql server datatypes like
`datetime`, `datetime2`, `datetimeoffset` are all treated as `datetime`
by graphql, so there is no 1:1 mapping for datetime datatypes. So we are
skipping supplying the DbType for these.
3. As explained in this stack overflow article
[here](https://stackoverflow.com/questions/6868300/net-guid-uppercase-string-format),
the GUID when converted to string, is output in lowercase. That's why
when the query to verify the MsSql test `InsertOneInSupportedTypes` is
written, we have explicitly cast `guid_types` to lower case.

## Additional change
1. The `real` sql server datatype was incorrectly mapped to `long`
dotnet type. It has been correctly mapped to `decimal` now.

## Issue with DateTime/DateTimeOffset Sql Server data types
For a mutation request like this, we get a response:
<img width="873" alt="image"
src="https://user-images.githubusercontent.com/34566234/235473783-87b810b1-bf10-44a6-8671-c54b74e44dac.png">

and the parameters are generated as:
<img width="510" alt="image"
src="https://user-images.githubusercontent.com/34566234/235470342-5331c755-abdc-4e05-9f9c-467e406d0038.png">

the parameter generated is of type `DateTimeOffset` (as evident from the
value highlighted) but the underlying Sql Server datatype for the
`datetime_types` field is `datetime`. Even though we supplied a value
`1999-01-08 09:20:00` in the mutation, it became `08-01-1999 09:20:00
+05:30`, because that's how graphql treats datetime/datetimoffset/date
types. Then this implicit conversion from a value of type DateTimeOffset
to DateTime cannot happen throwing an exception with message : `Failed
to convert parameter value from a DateTimeOffset to a DateTime.` Hence
we cannot add a mapping from `typeof(DateTime) = DbType.DateTime`.

Another option that came to my mind and would probably come to your mind
as well would be to have a mapping from`typeof(DateTime) =
DbType.DateTimeOffset`, since then there can be implicit conversion from
value of `DateTimeOffset` type to `DbType.DateTimeOffset`. But that has
its own flaws. Lets look at the below screenshots:

Consider this mutation:
<img width="885" alt="image"
src="https://user-images.githubusercontent.com/34566234/235477733-19c18d73-7eaa-42ba-9484-52984c62baad.png">

and the parameters generated for the query (mutation ran successfully
because the default value for the `instant` PK column was inserted into
the table):

<img width="425" alt="image"
src="https://user-images.githubusercontent.com/34566234/235478676-6c2b5bc7-80bc-497d-a53d-e3ecff80538e.png">
The parameter value is of sql server type datetime but the DbType is
populated as `DbType.DateTimeOffset`. This upcasting cannot happen
correctly and so we are returned with 0 records. So, this approach won't
work as well.

**Conclusion:** We won't populate DbType for these sql server types as
it is not required and would prevent data loss and any unforeseen
exceptions.

## How was this tested?

- [x] Integration Tests
- [x] Unit Tests - The unit tests that were skipped before saying that
bytearray datatype is not supported are now running successfully for
MsSql.

## Sample Request(s)
Request method: POST
Request URL: https://localhost:5001/api/SupportedType/
Request:
```json
{
    "bytearray_types": null
}
```
Response:
```json
{
    "value": [
        {
            "typeid": 5001,
            "byte_types": null,
            "short_types": null,
            "int_types": null,
            "long_types": null,
            "string_types": null,
            "single_types": null,
            "float_types": null,
            "decimal_types": null,
            "boolean_types": null,
            "date_types": null,
            "datetime_types": null,
            "datetime2_types": null,
            "datetimeoffset_types": null,
            "smalldatetime_types": null,
            "bytearray_types": null,
            "guid_types": "27742da6-99f2-4e58-9fd8-4f143aea46b0"
        }
    ]
}
```